requirements.txt
streamlit
pandas
plotly
openpyxl
sqlalchemy
google-genai